Conversion back and forth location ID and sector coords
This will be used for #22 and maybe other stuff.  In particular, with
this we don't really need to track the ship's position to update
buildings (though we probably want to do it anyway, for #15), and we
don't need to store coords any more, so we could shave a few bytes off
per building.