Hiding the storage use gauge on Firefox
To tell the user how much of the available storage space is being used
we need StorageArea's getBytesInUse, MAX_ITEMS, and QUOTA_BYTES, which
Firefox doesn't implement.  So we just hide the gauge.

If/when FF supports this, it should start showing the gauge without us
needing to change anything here.