Nav filter = base/planet trade filter
After using this in the wild, I feel that using the same "sticky filter"
for the building list in nav and the starbase/planet trade screen is
better.  When you set a filter for a trading run, you really want the
same filter whether you peek at the list from the nav screen or at the
planet; keeping separate persistent settings is actually more burden,
you end up having to set the same filter twice.

The filter in Overview -> Buildings is still independent.